Updated code for processing data for Pix4D
Removed test code. Updated existing code to reflect positive-negative switch for camera_offset, to work off of only those images that have been flagged as being ready for pre-Pix4D processing, and to update the DB to indicate the flyovers are ready for Pix4D processing when the imagery are copied over.